% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexscan.R
\name{choropleth}
\alias{choropleth}
\title{Display choropleth map}
\usage{
choropleth(polygons, fls, region_color = "#F0F0F0",
  rank = 1:length(fls$cluster), pval = 1, ...)
}
\arguments{
\item{polygons}{A SpatialPolygonsDataFrame.}

\item{fls}{An rflexscan object.}

\item{region_color}{Color of regions that are not included in any clusters.}

\item{rank}{An integer vector which specifies ranks of clusters to be displayed.}

\item{pval}{A threshold of P-value. Clusters with P-values of <\code{pval} will be displayed.}

\item{...}{Other parameters to be passed to plot function.}
}
\description{
Display choropleth map of detected clusters.
}
\details{
Clusters are colored using the current palette. Please use \link{palette}
function to specify colors of each cluster. Note that clusters with ranks
larger than the number of colors in the palette are not highlighted.
}
