% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexscan.R
\name{flexscan}
\alias{flexscan}
\title{Analyze spatial count data}
\usage{
flexscan(case, coordinates, adj_mat, name = row.names(case),
  clustersize = 15, radius = 6370, model = flexscan.model,
  stattype = flexscan.stattype, scanmethod = flexscan.scanmethod,
  ralpha = 0.2, cartesian = FALSE, simcount = 999,
  rantype = flexscan.rantype, ranseed = 4586111, comments = "")
}
\arguments{
\item{case}{The frequency of disease in each area.
The first column is the observed number of diseases, and the
second column is the expected number of diseases under the null
hypothesis, or the background population at risk in each area.}

\item{coordinates}{The geographic coordinates for each area.
Coordinates may be specified either using the standard Cartesian 
coordinates system or in latitude and longitude. The first column is the
x-coordinate or the latitude, and the second column is the y-coordinate or
longitude.}

\item{adj_mat}{The adjacency matrix.}

\item{name}{The name of each area.}

\item{clustersize}{The number of maximum spatial cluster size to scan.}

\item{radius}{Radius of Earth to calculate a distance between two sets of latitude and
longitude. It is approximately 6370 km in Japan.}

\item{model}{Statistical model to be used (case-insensitive).
\describe{
  \item{"POISSON"}{for the data of the 'observed number' and the 'expected number'}
  \item{"BINOMIAL"}{for the data of the 'observed number' and the 'population'}
}}

\item{stattype}{Statistic type to be used (case-insensitive).
\describe{
  \item{"ORIGINAL"}{the likelihood ratio statistic by Kulldorff and Nagarwalla (1995)}
  \item{"RESTRICTED"}{the restricted likelihood ratio statistic by Tango (2008), with a preset parameter \code{ralpha} for restriction}
}}

\item{scanmethod}{Scanning method to be used (case-insensitive).
\describe{
  \item{"FLEXIBLE"}{flexible scan statistic by Tango and Takahashi (2005)}
  \item{"CIRCULAR"}{circular scan statistic by Kulldorff}
}}

\item{ralpha}{Parameter for the restricted likelihood ratio statistic.}

\item{cartesian}{Set \code{TRUE} for Cartesian coordinates.}

\item{simcount}{The number of Monte Carlo replications to calculate a p-value for statistical
test.}

\item{rantype}{The type of random number for Monte Carlo simulation (case-insensitive).
\describe{
  \item{"MULTINOMIAL"}{Total number of cases in whole area is fixed. It can be chosen in either Poisson or Binomial model.}
  \item{"POISSON"}{Total number of cases is not fixed. It can be chosen in Poisson model.}
}}

\item{ranseed}{The seed for generating random numbers in the Monte Carlo simulation.}

\item{comments}{Comments for the analysis which will be written in summary.}
}
\value{
An \code{rflexscan} object which contains analysis results and specified
parameters.
}
\description{
Analyzes spatial count data using the flexible/circular 
spatial scan statistic.
}
\examples{
# load sample data
data("saitama_heart_m")
data("saitama_coordinates")
data("saitama_adj_mat")

# Run FleXScan with default parameters
fls <- flexscan(saitama_heart_m, saitama_coordinates, saitama_adj_mat)

# Print summary to the terminal
summary(fls)

# Plot graph
plot(fls)
labs <- 1:length(fls$cluster)
legend("topright", legend = labs, col = rainbow(length(fls$cluster)), lty = 1)

}
\references{
\itemize{
  \item Tango T. and Takahashi K. (2005). A flexibly shaped spatial scan statistic for detecting clusters, International Journal of Health Geographics 4:11.
  \item Kulldorff M. and Nagarwalla N. (1995). Spatial disease clusters: Detection and Inference. Statistics in Medicine 14:799-810.
  \item Tango T. (2008). A spatial scan statistic with a restricted likelihood ratio. Japanese Journal of Biometrics 29(2):75-95.
}
}
