% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexscan.R
\name{print.rflexscanCluster}
\alias{print.rflexscanCluster}
\title{Print rflexscanCluster object}
\usage{
\method{print}{rflexscanCluster}(x, ...)
}
\arguments{
\item{x}{An rflexscanCluster object to be printed.}

\item{...}{Ignored.}
}
\value{
No return value, called for side effects.
Prints details of a single detected cluster, including member regions and 
statistics.
}
\description{
Print method for the rflexscanCluster object.
}
