% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexscan.R
\docType{package}
\name{rflexscan-package}
\alias{rflexscan}
\alias{rflexscan-package}
\title{Analyze spatial count data using the flexible spatial scan statistic}
\description{
The rflexscan package provides functions and classes to analyze spatial count
data using the flexible spatial scan statistic developed by Tango and Takahashi (2005).
The original FleXScan software developed by Takahashi et al. is also 
available at \url{https://sites.google.com/site/flexscansoftware/home}.
}
\references{
\itemize{
  \item Tango T. and Takahashi K. (2005). A flexibly shaped spatial scan statistic for detecting clusters, International Journal of Health Geographics 4:11.
  \item Takahashi K, Yokoyama T and Tango T. (2010). FleXScan v3.1: Software for the Flexible Scan Statistic. National Institute of Public Health, Japan.
}
}
