% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexscan.R
\name{rflexscan}
\alias{rflexscan}
\title{Detect spatial disease clusters using the flexible/circular scan statistic}
\usage{
rflexscan(
  x,
  y,
  lat,
  lon,
  name,
  observed,
  expected,
  population,
  nb,
  clustersize = 15,
  radius = 6370,
  stattype = "ORIGINAL",
  scanmethod = "FLEXIBLE",
  ralpha = 0.2,
  simcount = 999,
  rantype = "MULTINOMIAL",
  comments = "",
  verbose = FALSE,
  secondary = NULL,
  clustertype = "HOT"
)
}
\arguments{
\item{x}{A vector of X-coordinates.}

\item{y}{A vector of Y-coordinates.}

\item{lat}{(DEPRECATED) A vector of latitude.}

\item{lon}{(DEPRECATED) A vector of longitude.}

\item{name}{A vector of names of each area.}

\item{observed}{A vector with the observed number of disease cases.}

\item{expected}{A vector with the expected number of disease cases under the null hypothesis. 
This is used on "Poisson" model.}

\item{population}{A vector with the background population at risk in each area. 
This is used on "Binomial" model.}

\item{nb}{A neighbors list or an adjacency matrix.}

\item{clustersize}{The number of maximum spatial cluster size to scan, i.e., the maximum number 
of regions included in the detected cluster}

\item{radius}{Radius of Earth to calculate a distance between two sets of latitude and
longitude. It is approximately 6370 km in Japan. This parameter is used when 
\code{lat} and \code{lon} are specified. This is DEPRECATED. The 
distance calculated using this parameter is not accurate. This feature is 
implemented to maintain compatibility with FleXScan. It is recommended to 
transform latitude and longitude onto the Cartesian coordinate system 
beforehand and use the \code{x} and \code{y} parameters that are projected
coordinates.}

\item{stattype}{Statistic type to be used (case-insensitive).
\describe{
  \item{"ORIGINAL"}{the likelihood ratio statistic by Kulldorff and
  Nagarwalla (1995)}
  \item{"RESTRICTED"}{the restricted likelihood ratio statistic by Tango 
  (2008), with a preset parameter \code{ralpha} for restriction}
}}

\item{scanmethod}{Scanning method to be used (case-insensitive).
\describe{
  \item{"FLEXIBLE"}{flexible scan statistic by Tango and Takahashi (2005)}
  \item{"CIRCULAR"}{circular scan statistic by Kulldorff (1997)}
}}

\item{ralpha}{Threshold parameter of the middle p-value for the restricted likelihood ratio
statistic.}

\item{simcount}{The number of Monte Carlo replications to calculate a p-value for statistical
test.}

\item{rantype}{The type of random number for Monte Carlo simulation (case-insensitive).
\describe{
  \item{"MULTINOMIAL"}{Total number of cases in whole area is fixed. It can 
  be chosen in either Poisson or Binomial model.}
  \item{"POISSON"}{Total number of cases is not fixed. It can be chosen in 
  Poisson model.}
}}

\item{comments}{Comments for the analysis which will be written in summary.}

\item{verbose}{Print progress messages.}

\item{secondary}{The number of secondary clusters to be enumerated. If \code{NULL} is 
specified (default), the search for secondary clusters is stopped when the 
Monte Carlo p-value reaches 1.}

\item{clustertype}{Type of cluster to be scanned.
\describe{
  \item{"HOT"}{Hot-spot clusters with elevated risk.}
  \item{"COLD"}{Cold-spot clusters with reduced risk.}
  \item{"BOTH"}{Hot- and cold-spot clusters simultaneously.}
}}
}
\value{
An \code{rflexscan} object which contains analysis results and specified
parameters.
}
\description{
This function analyzes spatial count data using the flexible spatial scan 
statistic developed by Tango and Takahashi (2005) or Kulldorff's circular 
spatial scan statistic (1997), and detect spatial disease clusters.
}
\details{
Centroid coordinates for each region should be specified EITHER by Cartesian 
coordinates using arguments \code{x} and \code{y} or by latitudes and 
longitudes using arguments \code{lat} and \code{lon}.
Note that \code{lat} and \code{lon} are DEPRECATED due to accuracy issues.
This feature is implemented to maintain compatibility with FleXScan software.
We recommend to transform latitude and longitude onto the Cartesian 
coordinate system beforehand (using \code{spTransform} function in sp package,
for example) and use the \code{x} and \code{y} parameters that are projected
coordinates.
}
\examples{
# load sample data (North Carolina SIDS data)
library(spdep)
data("nc.sids")

# calculate the expected numbers of cases
expected <- nc.sids$BIR74 * sum(nc.sids$SID74) / sum(nc.sids$BIR74)

# run FleXScan
fls <- rflexscan(x = nc.sids$x, y = nc.sids$y,
                 observed = nc.sids$SID74,
                 expected = expected,
                 name = rownames(nc.sids),
                 clustersize = 10,
                 nb = ncCR85.nb)

# print rflexscan object
print(fls)

# print properties of the most likely cluster
print(fls$cluster[[1]])

# print summary to the terminal
summary(fls)

# plot graph
plot(fls, col = palette())
labs <- 1:length(fls$cluster)
legend("bottomleft", legend = labs, col = palette(), lty = 1)

}
\references{
Otani T. and Takahashi K. (2021). Flexible scan statistics for 
  detecting spatial disease clusters: The rflexscan R package, Journal of 
  Statistical Software 99:13.
  
  Tango T. and Takahashi K. (2005). A flexibly shaped spatial scan 
  statistic for detecting clusters, International Journal of Health 
  Geographics 4:11.
  
  Kulldorff M. and Nagarwalla N. (1995). Spatial disease clusters: 
  Detection and Inference. Statistics in Medicine 14:799-810.
  
  Kulldorff M. (1997). A spatial scan statistic. Communications in 
  Statistics: Theory and Methods, 26:1481-1496.
  
  Tango T. (2008). A spatial scan statistic with a restricted 
  likelihood ratio. Japanese Journal of Biometrics 29(2):75-95.
}
\seealso{
\link{summary.rflexscan}, \link{plot.rflexscan}, \link{choropleth}
}
