% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runFleXScan}
\alias{runFleXScan}
\title{Run main routine of FleXScan.}
\usage{
runFleXScan(setting, case_mat, coord_mat, adj_mat)
}
\arguments{
\item{setting}{A list of parameter setting.}

\item{case_mat}{A matrix of case counts.}

\item{coord_mat}{A matrix of coordinates.}

\item{adj_mat}{A matrix of neighbourhood relationships.}
}
\value{
A list containing the detected clusters, test statistics, and Monte Carlo p-values.
}
\description{
Run main routine of FleXScan.
}
