% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexscan.R
\name{summary.rflexscan}
\alias{summary.rflexscan}
\title{Summarizing rflexscan results}
\usage{
\method{summary}{rflexscan}(object, ...)
}
\arguments{
\item{object}{An rflexscan object to be summarized.}

\item{...}{Ignored.}
}
\description{
Summary method for rflexscan objects.
}
\seealso{
\link{rflexscan}
}
