% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexscan.R
\name{plot.rflexscan}
\alias{plot.rflexscan}
\title{Graph plotting of flexscan results}
\usage{
\method{plot}{rflexscan}(object, col = rainbow(length(object$cluster)),
  rank = 1:length(object$cluster), pval = 1,
  vertexsize = max(object$coordinates[, 1]) - min(object$coordinates[,
  1]), xlab = colnames(object$coordinates)[1],
  ylab = colnames(object$coordinates)[2],
  xlim = c(min(object$coordinates[, 1]), max(object$coordinates[, 1])),
  ylim = c(min(object$coordinates[, 2]), max(object$coordinates[, 2])),
  ...)
}
\description{
Graph plotting of flexscan results
}
\examples{
# highlight all clusters
plot(flexscan.output)

# highlight clusters with rank 1 and 2
plot(flexscan.output, rank = c(1,2))

# highlight clusters of P-value <= 0.05
plot(flexscan.output, pval = 0.05)

}
