% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexscan.R
\name{plot.rflexscan}
\alias{plot.rflexscan}
\title{Graph plotting of flexscan results}
\usage{
\method{plot}{rflexscan}(x, rank = 1:length(x$cluster), pval = 1,
  vertexsize = max(x$coordinates[, 1]) - min(x$coordinates[, 1]),
  xlab = colnames(x$coordinates)[1], ylab = colnames(x$coordinates)[2],
  xlim = c(min(x$coordinates[, 1]), max(x$coordinates[, 1])),
  ylim = c(min(x$coordinates[, 2]), max(x$coordinates[, 2])), ...)
}
\arguments{
\item{x}{An rflexscan object.}

\item{rank}{An integer vector which specifies ranks of clusters to be displayed.}

\item{pval}{A threshold of P-value. Clusters with P-values of <\code{pval} will be displayed.}

\item{vertexsize}{Size of vertex of the graph.}

\item{xlab}{A label of the x axis.}

\item{ylab}{A label of the y axis.}

\item{xlim}{The x limits of the plot.}

\item{ylim}{The y limits of the plot.}

\item{...}{Other parameters to be passed to \link{plot.igraph} function.}
}
\description{
Display detected clusters by a graph representation.
}
\details{
Clusters are colored using the current palette. Please use \link{palette}
function to specify colors of each cluster. Note that clusters with ranks
larger than the number of colors in the palette are not highlighted.
}
\examples{
\dontrun{
# display all clusters
plot(rflexscan_object)

# display clusters with rank 1 and 2
plot(rflexscan_object, rank = c(1,2))

# display clusters of P-value <= 0.05
plot(rflexscan_object, pval = 0.05)
}

}
