% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexscan.R
\name{print.rflexscan}
\alias{print.rflexscan}
\title{Print rflexscan object}
\usage{
\method{print}{rflexscan}(x, ...)
}
\arguments{
\item{x}{An rflexscan object to be printed.}

\item{...}{Ignored.}
}
\description{
Print method for the rflexscan object.
}
\seealso{
\link{rflexscan}
}
