% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexscan.R
\name{print.summary.rflexscan}
\alias{print.summary.rflexscan}
\title{Print summary of flexscan results}
\usage{
\method{print}{summary.rflexscan}(x, ...)
}
\arguments{
\item{x}{An summary.rflexscan object to be printed.}

\item{...}{Ignored.}
}
\value{
No return value, called for side effects.
Prints the summarized results of the flexible spatial scan statistic, 
including the number of clusters and test statistics.
}
\description{
Print summary of flexscan results to the terminal.
}
\seealso{
\link{rflexscan}, \link{summary.rflexscan}
}
