% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexscan.R
\name{summary.rflexscan}
\alias{summary.rflexscan}
\title{Summarizing rflexscan results}
\usage{
\method{summary}{rflexscan}(object, ...)
}
\arguments{
\item{object}{An rflexscan object to be summarized.}

\item{...}{Ignored.}
}
\value{
An object of class \code{"summary.rflexscan"} including summary tables of 
detected clusters, test statistics, and p-values. This object can be printed 
with \code{\link{print.summary.rflexscan}}.
}
\description{
Summary method for rflexscan objects.
}
\seealso{
\link{rflexscan}
}
